% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{ctstanfit.fold}
\alias{ctstanfit.fold}
\title{Fit ctsem model to longitudinal dataset}
\usage{
ctstanfit.fold(train.dataset, ctmodel)
}
\arguments{
\item{train.dataset}{Longitudinal dataset.}

\item{ctmodel}{ctmodel.}
}
\value{
object of class ctStanFit.
}
\description{
Fit ctsem model to longitudinal dataset
}
