% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{cut.intervals}
\alias{cut.intervals}
\title{Split intervals so that no interval is longer than max.interval}
\usage{
\method{cut}{intervals}(id, tstart, interval.length, max.interval)
}
\arguments{
\item{id}{Vector of ids.}

\item{tstart}{Vector of start times}

\item{interval.length}{Vector of lengths of intervals}

\item{max.interval}{Maximum interval length in output}
}
\value{
data.table with colums id, cut.time that can be used to loop over ids calling survSplit.
}
\description{
Split intervals so that no interval is longer than max.interval
}
