% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{fit.poissontsplit}
\alias{fit.poissontsplit}
\title{Merge imputed biomarker values with survival data.table and fit Poisson regression model}
\usage{
fit.poissontsplit(
  kal.wide,
  dataSurv,
  timeinvar.surv,
  biomarkers,
  splines = FALSE,
  stan = TRUE,
  vb = TRUE
)
}
\arguments{
\item{kal.wide}{Data.table of imputed biomarker values, with time in column named tstart.}

\item{dataSurv}{Data.table of survival values with columns id, Time.cens, event}

\item{timeinvar.surv}{Character vector of names of time-invariant covariates in dataSurv.train}

\item{biomarkers}{Character vector of names of biomarkers in dataSurv.train.}

\item{splines}{Logical variable as to whether to model time as a spline function.}

\item{stan}{Logical variable whether to use Stan to fit the Poisson regression}

\item{vb}{Logical variable whether to use variational Bayes approximation rather than optimising.}
}
\value{
Matrix of regression coefficients with one column.
}
\description{
Merge imputed biomarker values with survival data.table and fit Poisson regression model
}
