% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{listmodels}
\alias{listmodels}
\title{Generate list of five stanct models: linear mixed model (LMM), autoregressive drift and Wiener diffusion process only, LMM + drift, LMM + diffusion, LMM + drift + diffusion}
\usage{
listmodels(biomarkers, timeinvar.long)
}
\arguments{
\item{biomarkers}{Character vector of names of biomarkers.}

\item{timeinvar.long}{Character vector of names of time-invariant covariates}
}
\value{
List of five objects of class ctStanModel.
}
\description{
Generate list of five stanct models: linear mixed model (LMM), autoregressive drift and Wiener diffusion process only, LMM + drift, LMM + diffusion, LMM + drift + diffusion
}
\examples{
data(pbc)
models.list <- listmodels(pbc$biomarkers, pbc$timeinvar.long)
summary(models.list)

}
