% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc-data.R
\docType{data}
\name{pbc}
\alias{pbc}
\title{Mayo clinic primary biliary cirrhosis dataset}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(pbc)
}
\description{
Mayo clinic primary biliary cirrhosis dataset
}
\examples{
data(pbc)
models.list <- listmodels(pbc$biomarkers, pbc$timeinvar.long) 
summary(models.list) 
}
\keyword{datasets}
