% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{predict.testdata}
\alias{predict.testdata}
\title{Predict event indicators from Poisson regression model using imputed biomarker values from kal.wide}
\usage{
\method{predict}{testdata}(beta, imputed.dt, dataSurv, timeinvar.surv, biomarkers, landmark.time)
}
\arguments{
\item{beta}{Matrix of regression coefficients with one column.}

\item{imputed.dt}{Data.table of imputed values generated by kalmanwide(), one record per person-time interval.}

\item{dataSurv}{Data.table with column id used to fit ctsem model.}

\item{timeinvar.surv}{Character vector of names of time-invariant covariates in dataSurv.train.}

\item{biomarkers}{Character vector of names of biomarkers in dataSurv.train.}

\item{landmark.time}{Landmark time for start of prediction, as numeric variable.}
}
\value{
Data.table with columns id, event, tstart, tstop, toobs, covariates, p.event, cumulative survival prob for each individual.
}
\description{
Predict event indicators from Poisson regression model using imputed biomarker values from kal.wide
}
