% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{summary.predictions}
\alias{summary.predictions}
\title{Summary statistics for predictive performance}
\usage{
\method{summary}{predictions}(testdata)
}
\arguments{
\item{testdata}{data.table with columns event, p.event, tobs}
}
\description{
Summary statistics for predictive performance
}
