% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{tabulate.predictions}
\alias{tabulate.predictions}
\title{Summary statistics for predictive performance}
\usage{
tabulate.predictions(testdata)
}
\arguments{
\item{testdata}{data.table with columns event, p.event, tobs.}
}
\description{
Summary statistics for predictive performance
}
