% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{test.imputed}
\alias{test.imputed}
\title{Predict event indicators from Poisson regression model using imputed biomarker values from kal.wide}
\usage{
test.imputed(
  poisson.glm.train,
  ids.test,
  kal.wide,
  dataSurv,
  landmark.time,
  timeinvar.surv,
  biomarkers
)
}
\arguments{
\item{poisson.glm.train}{Object of class glm generated by train.poissontsplit.}

\item{ids.test}{Vector of individual identifiers}

\item{kal.wide}{Data.table of imputed values generated by kalmanwide(), one record per person-time interval.}

\item{dataSurv}{Data.table with column id used to fit ctsem model.}

\item{timeinvar.surv}{Character vector of names of time-invariant covariates in dataSurv.train.}

\item{biomarkers}{Character vector of names of biomarkers in dataSurv.train.}
}
\value{
Data.table with columns id, event, tstart, tstop, toobs, covariates, p.event, cumulative survival prob for each individual.
}
\description{
Predict event indicators from Poisson regression model using imputed biomarker values from kal.wide
}
