% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{train.poissontsplit}
\alias{train.poissontsplit}
\title{Merge imputed biomarker values with survival table and fit Poisson regression model}
\usage{
train.poissontsplit(
  kal.wide,
  dataSurv.train,
  timeinvar.surv,
  biomarkers,
  splines = FALSE
)
}
\arguments{
\item{kal.wide}{Data.table of imputed biomarker values, with time in column named tstart.}

\item{dataSurv.train}{Data.table of survival values with columns id, Time.cens, event}

\item{timeinvar.surv}{Character vector of names of time-invariant covariates in dataSurv.train}

\item{biomarkers}{Character vector of names of biomarkers in dataSurv.train.}
}
\value{
Object of class glm containing fitted Poisson regression model.
}
\description{
Merge imputed biomarker values with survival table and fit Poisson regression model
}
