% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{trainsplit.surv}
\alias{trainsplit.surv}
\title{Test/train split of survival table, censoring follow-up at landmark.time for individuals in ids.test}
\usage{
trainsplit.surv(ids.test, dataSurv, landmark.time)
}
\arguments{
\item{ids.test}{Vector of individual identifiers.}

\item{dataSurv}{Data.table with columns id = individual identifier, Time.cens = time of exit as numeric, event=0-1 indicator for event status at exit.}

\item{landmark.time}{Landmark time at which to censor follow-up for individuals in ids.test.}
}
\value{
Survival dataset censored at landmark.time for all individuals in ids.test, with Time renamed as Time.cens.
}
\description{
Test/train split of survival table, censoring follow-up at landmark.time for individuals in ids.test
}
