% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmseqfunctions.R
\name{kalmanwide}
\alias{kalmanwide}
\title{Impute values at beginning of each person-time interval from 0 to maxtime from ctstanfit object using Kalman filter, and cast to wide format}
\usage{
kalmanwide(train.fit, timestep, maxtime = maxtime)
}
\arguments{
\item{train.fit}{Object of class ctstanfit.}

\item{timestep}{Interval length.}

\item{maxtime}{Maximum time up to which to generate imputations.}
}
\value{
Data.table of imputed values with time in column tstart.
}
\description{
Impute values at beginning of each person-time interval from 0 to maxtime from ctstanfit object using Kalman filter, and cast to wide format
}
